Redesign BetConfirmationSheet: date banner, runner/syndicate identity, preset wagers, cleaner odds
- Add gameDate to SelectedBet and thread odd.gameDt through all 6 SelectedBet call sites
- BetGameBanner: show formatted date (EEE, MMM d) between matchup and time; replace Away/Home labels with team abbrs; upgrade bet label font to .subheadline; add accent-colored 'x' suffix to odds price; remove redundant implied % from banner
- BetConfirmationSheet: replace IDs card with single-row syndicate + runner identity card (icon + name, fetched async); replace large free-text wager field with preset buttons (½, 1, 2, 3) + # custom input; change computedUnits to Double and PlacedBet.units to Double for 0.5 support

+++ unitleagueios/Models/PlacedBet.swift
@@ -7,7 +7,7 @@ struct PlacedBet: Codable, Identifiable {
     let side: String        // "Away" or "Home"
     let price: Double
     let points: Double?     // spread value or O/U total; nil for ML
-    let units: Int
+    let units: Double
     let awayAbbr: String
     let homeAbbr: String
     let gameTime: String?   // raw "HH:mm:ss" from API
